Add stateful Angular component in Kotlin
To implement a Kotlin component, we need to declare a "shim" in Angular
that extends the Kotlin class which in turn contains the logic we want.

Now, the Kotlin class can create state (with a custom wrapper around
angular's signal) and render that state in the component. Cool beans.

Unfortunately I think this means we're including angular/core twice,
once for the Kotlin/JS and once for the Typescript. :-/