Extract auth to Angular-managed service
It turns out that Angular only honors state for signals that it "sees"
in the class metadata. The magic only goes so far :)

If component template reads a signal through a Kotlin function, Angular
doesn't recognize the component's lifecycle depends on that signal's
value. Instead we need to declare all state in Angular TypeScript and
feed it into the Kotlin as necessary.

Along the way I began distinguishing between signals and writeable
signals, initially I was trying to match the Angular types but I backed
away from that. But I kept the read-only signal type.