Manage flows service state in Angular
Like with auth previously, we need to manage state for the flows service
in Angular. Otherwise, we don't get reactive state :-/

This also removes all semblance of read API from signals. It doesn't
seem to work, at all. So let's not pretend it does :-)

This seems to get us to a point where angular works in the local
emulator (whereas before it only worked through `ng serve`).