Move flows state into flows service
Converting Firebase flow bundles into their flows should be part of the
flows service: "just give me the flows". This lets us make flows part of
the Angular state as usual, for example to render the list of flows or
to render the flows as balances (in previous commit).

It also refactors the flow list to take a list of flows (conveniently
sourced from the flows service's flow signal we just added).