Let user toggle flow exclusion
Users can click a flow in the sidebar to toggle its exclusion from the
cash flow simulation. This lets us explore the effects with or without a
cash flow quickly.

Along the way a refactor:

- With previous work on balances, we can use regular computed signals
  rather than an effect in the App state for flow balances.

The 'output' directive is special, it can only be used in the field
initializer. It can't be called in the constructor body. This prevents
us from passing it to Kotlin (we can't refer to fields on 'this' until
we call super). This is why we need the extra signal indirection in the
flow list component: kotlin manages the signal, and typescript forwards
that signal to the output in an effect.

An alternate implementation could have an abstract method on the kotlin
class, implemented in the typescript subclass, which emits to the output
directly …… although, the need to set it in the constructor might
prevent that (can the kt ctor call a function defined in the ts body,
referencing a ts instance variable?).