Support MatDialog + datepicker + KotlinJS interop
We need to include luxon in Kotlin so that we can reference luxon
datetimes coming in from JS. This is needed for interop with JS where we
pass a kotlin LocalDate to JS as a Luxon DateTime, which the DatePicker
manipulates (writing a Luxon DateTime as well). This DateTime needs to
get sent back to Kotlin and converted to a LocalDate.

A more concise approach might require a DateAdapter for the date picker.
It doesn't know how to work with a KotlinX datetime, which I gather uses
the Joda library on the JS platform. So we'd need a Joda DateAdapter for
the material date picker to manipulate an object that could be sent back
to the kotlin layer "natively".